Extracted ImageAlternateText plugin.
