+++ src/imagealternatetext/ui/alternatetextformview.js
