+++ src/imageengine.js
@@ -10,7 +10,6 @@
 import Plugin from 'ckeditor5-core/src/plugin';
 import buildModelConverter from 'ckeditor5-engine/src/conversion/buildmodelconverter';
 import WidgetEngine from './widget/widgetengine';
-import ImageAlternateTextCommand from './imagealternatetextcommand.js';
 import { modelToViewImage, viewToModelImage, modelToViewSelection } from './converters';
 import { toImageWidget } from './utils';
 
@@ -60,8 +59,5 @@ export default class ImageEngine extends Plugin {
 
 		// Creates fake selection label if selection is placed around image widget.
 		editing.modelToView.on( 'selection', modelToViewSelection( editor.t ), { priority: 'lowest' } );
-
-		// Register ImageAlternateTextCommand.
-		editor.commands.set( 'imageAlternateText', new ImageAlternateTextCommand( editor ) );
 	}
 }
